Refactor admin UI services into domain modules
