+++ backend/apps/admin_ui/routers/api.py
@@ -3,26 +3,33 @@ from typing import Optional
 from fastapi import APIRouter, Query
 from fastapi.responses import JSONResponse
 
-from backend.apps.admin_ui import services
 from backend.apps.admin_ui.utils import parse_optional_int, status_filter
+from backend.apps.admin_ui.services.dashboard import dashboard_counts
+from backend.apps.admin_ui.services.recruiters import api_recruiters_payload
+from backend.apps.admin_ui.services.cities import (
+    api_cities_payload,
+    api_city_owners_payload,
+)
+from backend.apps.admin_ui.services.slots import api_slots_payload
+from backend.apps.admin_ui.services.templates import api_templates_payload
 
 router = APIRouter(prefix="/api", tags=["api"])
 
 
 @router.get("/health")
 async def api_health():
-    counts = await services.dashboard_counts()
+    counts = await dashboard_counts()
     return counts
 
 
 @router.get("/recruiters")
 async def api_recruiters():
-    return JSONResponse(await services.api_recruiters_payload())
+    return JSONResponse(await api_recruiters_payload())
 
 
 @router.get("/cities")
 async def api_cities():
-    return JSONResponse(await services.api_cities_payload())
+    return JSONResponse(await api_cities_payload())
 
 
 @router.get("/slots")
@@ -33,13 +40,13 @@ async def api_slots(
 ):
     recruiter = parse_optional_int(recruiter_id)
     status_norm = status_filter(status)
-    payload = await services.api_slots_payload(recruiter, status_norm, limit)
+    payload = await api_slots_payload(recruiter, status_norm, limit)
     return JSONResponse(payload)
 
 
 @router.get("/templates")
 async def api_templates(city_id: Optional[int] = None, key: Optional[str] = None):
-    payload = await services.api_templates_payload(city_id, key)
+    payload = await api_templates_payload(city_id, key)
     status_code = 200
     if isinstance(payload, dict) and payload.get("found") is False:
         status_code = 404
@@ -65,6 +72,6 @@ async def api_template_keys():
 
 @router.get("/city_owners")
 async def api_city_owners():
-    payload = await services.api_city_owners_payload()
+    payload = await api_city_owners_payload()
     status_code = 200 if payload.get("ok") else 400
     return JSONResponse(payload, status_code=status_code)
