+++ backend/apps/admin_ui/routers/recruiters.py
@@ -4,20 +4,28 @@ from fastapi import APIRouter, Form, Request
 from fastapi.responses import HTMLResponse, RedirectResponse
 
 from backend.apps.admin_ui.config import templates
-from backend.apps.admin_ui import services
+from backend.apps.admin_ui.services.cities import list_cities
+from backend.apps.admin_ui.services.recruiters import (
+    build_recruiter_payload,
+    create_recruiter,
+    delete_recruiter,
+    get_recruiter_detail,
+    list_recruiters,
+    update_recruiter,
+)
 
 router = APIRouter(prefix="/recruiters", tags=["recruiters"])
 
 
 @router.get("", response_class=HTMLResponse)
 async def recruiters_list(request: Request):
-    recruiters = await services.list_recruiters()
+    recruiters = await list_recruiters()
     return templates.TemplateResponse("recruiters_list.html", {"request": request, "recruiters": recruiters})
 
 
 @router.get("/new", response_class=HTMLResponse)
 async def recruiters_new(request: Request):
-    cities = await services.list_cities()
+    cities = await list_cities()
     return templates.TemplateResponse("recruiters_new.html", {"request": request, "cities": cities})
 
 
@@ -30,20 +38,20 @@ async def recruiters_create(
     active: Optional[str] = Form(None),
     cities: Optional[List[str]] = Form(None),
 ):
-    payload: Dict[str, object] = services.build_recruiter_payload(
+    payload: Dict[str, object] = build_recruiter_payload(
         name=name,
         tz=tz,
         telemost=telemost,
         tg_chat_id=tg_chat_id,
         active=active,
     )
-    await services.create_recruiter(payload, cities=cities)
+    await create_recruiter(payload, cities=cities)
     return RedirectResponse(url="/recruiters", status_code=303)
 
 
 @router.get("/{rec_id}/edit", response_class=HTMLResponse)
 async def recruiters_edit(request: Request, rec_id: int):
-    data = await services.get_recruiter_detail(rec_id)
+    data = await get_recruiter_detail(rec_id)
     if not data:
         return RedirectResponse(url="/recruiters", status_code=303)
     return templates.TemplateResponse("recruiters_edit.html", {"request": request, **data})
@@ -59,18 +67,18 @@ async def recruiters_update(
     active: Optional[str] = Form(None),
     cities: Optional[List[str]] = Form(None),
 ):
-    payload: Dict[str, object] = services.build_recruiter_payload(
+    payload: Dict[str, object] = build_recruiter_payload(
         name=name,
         tz=tz,
         telemost=telemost,
         tg_chat_id=tg_chat_id,
         active=active,
     )
-    await services.update_recruiter(rec_id, payload, cities=cities)
+    await update_recruiter(rec_id, payload, cities=cities)
     return RedirectResponse(url="/recruiters", status_code=303)
 
 
 @router.post("/{rec_id}/delete")
 async def recruiters_delete(rec_id: int):
-    await services.delete_recruiter(rec_id)
+    await delete_recruiter(rec_id)
     return RedirectResponse(url="/recruiters", status_code=303)
