+++ backend/apps/admin_ui/services/__init__.py
@@ -1,0 +1,78 @@
+from .dashboard import dashboard_counts
+from .recruiters import (
+    api_recruiters_payload,
+    build_recruiter_payload,
+    create_recruiter,
+    delete_recruiter,
+    get_recruiter_detail,
+    list_recruiters,
+    update_recruiter,
+)
+from .cities import (
+    api_cities_payload,
+    api_city_owners_payload,
+    assign_city_owner,
+    city_owner_field_name,
+    create_city,
+    list_cities,
+    update_city_settings,
+)
+from .templates import (
+    api_templates_payload,
+    create_template,
+    delete_template,
+    get_stage_templates,
+    get_template,
+    list_templates,
+    stage_payload_for_ui,
+    templates_overview,
+    update_template,
+    update_templates_for_city,
+)
+from .slots import (
+    api_slots_payload,
+    create_slot,
+    list_slots,
+    recruiters_for_slot_form,
+)
+from .questions import (
+    get_test_question_detail,
+    list_test_questions,
+    update_test_question,
+)
+
+__all__ = [
+    "dashboard_counts",
+    "list_recruiters",
+    "create_recruiter",
+    "get_recruiter_detail",
+    "update_recruiter",
+    "delete_recruiter",
+    "build_recruiter_payload",
+    "list_cities",
+    "create_city",
+    "assign_city_owner",
+    "update_city_settings",
+    "city_owner_field_name",
+    "get_stage_templates",
+    "stage_payload_for_ui",
+    "templates_overview",
+    "update_templates_for_city",
+    "list_templates",
+    "create_template",
+    "get_template",
+    "update_template",
+    "delete_template",
+    "api_recruiters_payload",
+    "api_cities_payload",
+    "api_slots_payload",
+    "api_templates_payload",
+    "api_city_owners_payload",
+    "list_slots",
+    "recruiters_for_slot_form",
+    "create_slot",
+    "list_test_questions",
+    "get_test_question_detail",
+    "update_test_question",
+]
+
